+++ build-plugins/build-support/src/main/kotlin/aws/sdk/kotlin/gradle/dsl/CodeStyle.kt
@@ -40,7 +40,7 @@ fun Project.configureLinting(lintPaths: List<String>) {
                 attribute(Bundling.BUNDLING_ATTRIBUTE, objects.named(Bundling.SHADOWED))
             }
         }
-        ktlint("aws.sdk.kotlin.gradle:ktlint-rules:$repoToolsVersion")
+        ktlint("aws.sdk.kotlin.gradle:style:$repoToolsVersion")
     }
 
     // add the buildscript classpath which should pick up our custom ktlint-rules (via runtimeOnly dep on this plugin)
