+++ build-plugins/build-support/src/main/kotlin/aws/sdk/kotlin/gradle/dsl/MinorVersionStrategy.kt
@@ -7,9 +7,11 @@ package aws.sdk.kotlin.gradle.dsl
 
 import aws.sdk.kotlin.gradle.util.verifyRootProject
 import org.gradle.api.Project
+import org.gradle.api.artifacts.VersionCatalogsExtension
 import org.gradle.api.attributes.Bundling
 import org.gradle.api.tasks.JavaExec
 import org.gradle.kotlin.dsl.*
+import org.gradle.kotlin.dsl.getByType
 import org.gradle.language.base.plugins.LifecycleBasePlugin
 
 /**
@@ -24,6 +26,13 @@ fun Project.configureMinorVersionStrategyRules(lintPaths: List<String>) {
         ?.readText()
         ?: error("Missing ktlint-version.txt")
 
+    val repoToolsVersion = extensions
+        .getByType<VersionCatalogsExtension>()
+        .named("libs")
+        .findVersion("aws-kotlin-repo-tools-version")
+        .get()
+        .requiredVersion
+
     val minorVersionBumpKtlint by configurations.creating
 
     dependencies {
@@ -32,7 +41,7 @@ fun Project.configureMinorVersionStrategyRules(lintPaths: List<String>) {
                 attribute(Bundling.BUNDLING_ATTRIBUTE, objects.named(Bundling.SHADOWED))
             }
         }
-        minorVersionBumpKtlint(project(":ktlint-rules:minor-version-strategy"))
+        minorVersionBumpKtlint("aws.sdk.kotlin.gradle:minor-version-rules:$repoToolsVersion")
     }
 
     tasks.register<JavaExec>("verifyMinorVersionBump") {
@@ -41,14 +50,5 @@ fun Project.configureMinorVersionStrategyRules(lintPaths: List<String>) {
         classpath = minorVersionBumpKtlint
         mainClass.set("com.pinterest.ktlint.Main")
         args = lintPaths
-    }
-
-    tasks.register<JavaExec>("fixMinorVersionBump") {
-        group = LifecycleBasePlugin.VERIFICATION_GROUP
-        description = "Auto fix minor version bump rules"
-        classpath = minorVersionBumpKtlint
-        mainClass.set("com.pinterest.ktlint.Main")
-        args = listOf("-F") + lintPaths
-        jvmArgs("--add-opens=java.base/java.lang=ALL-UNNAMED")
     }
 }
