+++ ktlint/minor-version-rules/src/main/resources/META-INF/services/com.pinterest.ktlint.cli.ruleset.core.api.RuleSetProviderV3
@@ -1,4 +1,4 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-software.aws.ktlint.rules.CustomRuleSetProvider
+software.aws.ktlint.rules.MinorVersionRuleSetProvider
